requirements.txt
dm_haiku==0.0.10
jax==0.4.14
matplotlib==3.7.3
numpy==1.25.2
pandas==2.1.1
scipy==1.11.3
setuptools==65.5.0
